+++ src/DataObjects/CheckoutSessionStatusData.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Subster\PhpSdk\DataObjects;
 
 use Subster\PhpSdk\Concerns\Data;
+use Subster\PhpSdk\Enums\CheckoutPaymentState;
 use Subster\PhpSdk\Enums\CheckoutSessionStatus;
 use Subster\PhpSdk\Enums\WebhookEndpointEvent;
 
@@ -18,6 +19,8 @@ class CheckoutSessionStatusData extends Data
         public readonly CheckoutSessionStatus $status,
         public readonly ?WebhookEndpointEvent $event,
         public readonly ?array $data,
+        public readonly CheckoutPaymentState $payment_state = CheckoutPaymentState::RequiresPayment,
+        public readonly ?string $checkout_url = null,
     ) {}
 
     public static function fromSaloon(array $response): self
@@ -27,6 +30,12 @@ class CheckoutSessionStatusData extends Data
             status: CheckoutSessionStatus::from(strval($response['status'])),
             event: isset($response['event']) ? WebhookEndpointEvent::from(strval($response['event'])) : null,
             data: isset($response['data']) && is_array($response['data']) ? $response['data'] : null,
+            payment_state: isset($response['payment_state'])
+                ? CheckoutPaymentState::from(strval($response['payment_state']))
+                : ($response['status'] === CheckoutSessionStatus::Completed->value
+                    ? CheckoutPaymentState::Paid
+                    : CheckoutPaymentState::RequiresPayment),
+            checkout_url: isset($response['checkout_url']) ? strval($response['checkout_url']) : null,
         );
     }
 }