+++ src/Enums/BillingPortalFlow.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Subster\PhpSdk\Enums;
+
+enum BillingPortalFlow: string
+{
+    case Overview = 'overview';
+    case PaymentRecovery = 'payment_recovery';
+}