+++ src/Enums/PaymentStrategy.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Subster\PhpSdk\Enums;
+
+enum PaymentStrategy: string
+{
+    case HostedCheckout = 'hosted_checkout';
+    case DefaultThenCheckout = 'default_then_checkout';
+}